split block file
